[Parser] Various merge adjustments to account for scenarios which pass parsing but are not expected * String fields now error if they are not actually strings * 'model' is now a number directly instead of string * 'npcID' no longer needs special negative value handling for 'headerID' overlap * Adjusted handling of 'long' type field merging * null value merges now warn if they remove an existing field, and otherwise skip merge logic * Added type conversion warning for most numeric-array fields * Prevent some type conversion handling warnings with simpler internal logic * Added a debug warn for Coord X/Y values which do not use decimals (2900+ right now)

+++ .contrib/Source Code/Parser/FieldTypes/Providers.cs
@@ -254,7 +254,7 @@ namespace ATT.FieldTypes
                                         && objectData.ObjectType == "questID")
                                     {
                                         // we will use 'qis' as a way to know that the itemID can be cached directly to that quest instead of as an item cost
-                                        Objects.Merge(_data, "qis", itemID);
+                                        Objects.Merge(_data, "qis", (long)itemID);
                                         Remove("i", itemID);
                                     }
                                 }
